Implemented Map Class
Child class of graph.
The class currently builds a 2D array that will represent the top-down
digital parkour map. Each (x, y) value on the map points back to a node
in the graph's node container. I'm hoping that this will allow the map
to seamlessly integrate with the graph and be searched using the
shortest-path algorithm.